fix: git rebase with master and refactoring moved the banner under components to be generally accessible
