+++ src/discussions/posts/post/messages.js
@@ -11,7 +11,7 @@ const messages = defineMessages({
   },
   contentReported: {
     id: 'discussions.post.contentReported',
-    defaultMessage: 'Content reported for staff review',
+    defaultMessage: 'Content reported for staff to review',
   },
   follow: {
     id: 'discussions.post.follow',
